HSD #15014165241: fpga: altera-cvp: agilex5 support
This change is to support Agilex5 new register mapping, whereby
the only difference is at the CVP Credit register which is now
0x5C instead of offset 0x49.
Added a new field device_family_type in order to support the
current change and any other future variations in the CVP
device. The device type is checked against the VSEC start offset
address and VSEC length. It is 0x60 for Agilex5.

Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>