HSD #18037244108: net: phy: marvell: fix 88e1510 downshift counter errata
88e1510 PHY has an errata that causes the phy downshift counter
not cleared on a link powering down/up. This can cause the gigabit
link to downshift to a lower speed(100Mbps) intermittently.
Disabling and reenabling the downshift feature will clear the
downshift counter and this helps PHY to retry the gigabit link
upto programmed retry count times before downshifting. Currently
observed only for the copper link.

Signed-off-by: Rohan G Thomas <rohan.g.thomas@intel.com>