HSD #15015934876: dts: socfpga_agilex5.dtsi: Fix the clk-csr for XGMAC
This patch is to to set the clk-csr divider value. The value is
set based on the clk_csr_i clock speed and currently the hardware
is connected with the l4_sp_clk (100MHz) thus the clk-csr divider
need to choose the divider with the value 62, due to that, the
csr-clk value need to choose divider option 0.

Signed-off-by: Boon Khai Ng <boon.khai.ng@intel.com>