HSD# 16027100077-3: mtd: spi-nor: core: Check for null pointer to avoid soft error
Ensure that the pointer passed to module_put() in spi_nor_put_device()
is not NULL before use. In certain configurations, the pointer may be
uninitialized or NULL, which can lead to a null pointer dereference
and a soft kernel error.

This change adds a guard clause to return early, preventing the error
and maintaining system stability without requiring a reboot.

Signed-off-by: Khairul Anuar Romli <khairul.anuar.romli@altera.com>