HSD #16028190682: net: stmmac: dwmac1000: Disable flushing frames
Disable flushing frames when receive buffer is unavailable
in Rx Store and Forward mode. With flushing frames when
receive buffer is unavailable we are observing MTL Rx FIFO
is going out of sync with one frame always buffered in the
Rx FIFO and gmacgrp_debug reg has a value of 0x00000120
always, which implies Reading frame data.

By setting disable flushing frames on receive buffer
unavailable the issue is not reproducible. But in such
cases Rx DMA will go to suspend state on receive buffer
unavailable and we need to write to receive_poll_demand
register after adding descriptors. This is used to wake up
the Rx DMA from suspend state.

Signed-off-by: Rohan G Thomas <rohan.g.thomas@altera.com>